requirements.txt
mastodon.py
matplotlib
pandas
sqlalchemy
vader-multi
numpy